refactor: Split _utils module into a _utils subpackage
