+++ src/apify_client/_resource_clients/user.py
@@ -16,7 +16,7 @@ from apify_client._models import (
     UserPublicInfo,
 )
 from apify_client._resource_clients._resource_client import ResourceClient, ResourceClientAsync
-from apify_client._utils import response_to_dict
+from apify_client._utils.http import response_to_dict
 
 if TYPE_CHECKING:
     from apify_client.types import Timeout
