+++ src/apify_client/_resource_clients/webhook.py
@@ -14,7 +14,7 @@ from apify_client._models import (
     WebhookUpdate,
 )
 from apify_client._resource_clients._resource_client import ResourceClient, ResourceClientAsync
-from apify_client._utils import response_to_dict
+from apify_client._utils.http import response_to_dict
 
 if TYPE_CHECKING:
     from apify_client._literals import WebhookEventType
